backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-google-genai>=0.0.1
sentence-transformers>=2.2.2
huggingface-hub>=0.4.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.2